Разбить написанный ранее js код на отдельные модули
