+++ index.html
@@ -250,9 +250,7 @@
     </section>
   </template>
 
-  <script type="text/javascript" src="js/functions.js"></script>
-  <script type="text/javascript" src="js/main.js"></script>
-
+  <script type="module" src="js/main.js"></script>
 </body>
 
 </html>
